Reversing the multiple threads

+++ Apps/SysinMain.cpp
@@ -4,7 +4,7 @@ int main(int argC, char** argV) {
     int result = Commons::POSIXErrors::OK;
 
     // Main process
-    result = std::max(SysinMainCaller::start(argC, argV), result);
+    result = std::max((int)SysinMainCaller::start(argC, argV), result);
 
     return result;
 }
